Refactor server files
